Expose a bad payload handler for more consistent behavior.
